+++ CourseLearning.Desktop/App.xaml
@@ -1,0 +1,7 @@
+﻿<Application x:Class="CourseLearning.Desktop.App"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             StartupUri="MainWindow.xaml">
+	<Application.Resources>
+	</Application.Resources>
+</Application>
