+++ CourseLearning.Desktop/App.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows;
+
+namespace CourseLearning.Desktop
+{
+    /// <summary>
+    /// Логика взаимодействия для App.xaml
+    /// </summary>
+    public partial class App : Application
+    {
+    }
+}
