+++ CourseLearning.Desktop/CourseLearning.Desktop.csproj
@@ -1,0 +1,33 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>WinExe</OutputType>
+    <TargetFramework>net7.0-windows</TargetFramework>
+    <Nullable>enable</Nullable>
+    <UseWPF>true</UseWPF>
+
+    <RootNamespace>CourseLearning.Desktop</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Remove="Pages\ReadingCoursesPage.xaml.cs~RF242e582.TMP" />
+    <None Remove="Resources\Icons\CourseLearning.ico" />
+    <None Remove="Resources\Icons\icons8-courses-48.png" />
+    <None Remove="Resources\Icons\icons8-help-48.png" />
+    <None Remove="Resources\Icons\icons8-list-courses-48.png" />
+    <None Remove="Resources\Icons\icons8-logout-48.png" />
+    <None Remove="Resources\Icons\icons8-profile.png" />
+    <None Remove="Resources\Icons\icons8-writing-48.png" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Resource Include="Resources\Icons\CourseLearning.ico" />
+    <Resource Include="Resources\Icons\icons8-courses-48.png" />
+    <Resource Include="Resources\Icons\icons8-help-48.png" />
+    <Resource Include="Resources\Icons\icons8-list-courses-48.png" />
+    <Resource Include="Resources\Icons\icons8-logout-48.png" />
+    <Resource Include="Resources\Icons\icons8-profile.png" />
+    <Resource Include="Resources\Icons\icons8-writing-48.png" />
+  </ItemGroup>
+
+</Project>
