+++ CourseLearning.Desktop/MainWindow.xaml
@@ -1,0 +1,78 @@
+﻿<Window x:Class="CourseLearning.Desktop.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        mc:Ignorable="d"
+        MinHeight="600"
+        MinWidth="1100"
+        Background="AliceBlue"
+        Title="CourseLearning - Lite"
+        WindowStartupLocation="CenterScreen">
+	<Grid>
+		<Grid.ColumnDefinitions>
+			<ColumnDefinition Width="Auto" />
+			<ColumnDefinition />
+		</Grid.ColumnDefinitions>
+
+		<!-- Sidebar -->
+		<Grid Background="LightSkyBlue">
+			<StackPanel Grid.Column="0" Width="200" VerticalAlignment="Center">
+				<!-- <Button Content="Profile" Margin="10" Background="LightSkyBlue" /> -->
+
+				<Grid Margin="0 0 0 0">
+					<Ellipse HorizontalAlignment="Left" Height="35" Margin="10,0,0,0" VerticalAlignment="Center" Width="35">
+						<Ellipse.Fill>
+							<ImageBrush ImageSource="/Resources/Icons/icons8-writing-48.png" Stretch="Fill" />
+						</Ellipse.Fill>
+					</Ellipse>
+					<Button x:Name="sideBarCreatingCourses" Margin="50,10,0,10" HorizontalAlignment="Left" Width="140"
+                    Background="AliceBlue" Click="sideBarCreatingCourses_Click">
+						<TextBlock Text="Создание курсов" TextWrapping="Wrap" TextAlignment="Center" FontSize="16" />
+						<Button.Resources>
+							<Style TargetType="Border">
+								<Setter Property="CornerRadius" Value="10"></Setter>
+							</Style>
+						</Button.Resources>
+					</Button>
+				</Grid>
+
+				<Grid Margin="0 10 0 10">
+					<Ellipse HorizontalAlignment="Left" Height="35" Margin="10,0,0,0" VerticalAlignment="Center" Width="35">
+						<Ellipse.Fill>
+							<ImageBrush ImageSource="/Resources/Icons/icons8-courses-48.png" Stretch="Fill" />
+						</Ellipse.Fill>
+					</Ellipse>
+					<Button x:Name="sideBarReadingCourses" Margin="50,10,0,10" HorizontalAlignment="Left" Width="140"
+                    Background="AliceBlue" Click="sideBarReadingCourses_Click">
+						<TextBlock Text="Прохождение курсов" TextWrapping="Wrap" TextAlignment="Center" FontSize="16" />
+						<Button.Resources>
+							<Style TargetType="Border">
+								<Setter Property="CornerRadius" Value="10"></Setter>
+							</Style>
+						</Button.Resources>
+					</Button>
+				</Grid>
+
+				<Grid Margin="0 0 0 0">
+					<Ellipse HorizontalAlignment="Left" Height="35" Margin="10,0,0,0" VerticalAlignment="Center" Width="35">
+						<Ellipse.Fill>
+							<ImageBrush ImageSource="/Resources/Icons/icons8-help-48.png" Stretch="Fill" />
+						</Ellipse.Fill>
+					</Ellipse>
+					<Button x:Name="sideBarHelp" Margin="50,10,0,10" HorizontalAlignment="Left" Width="140"
+                    Background="AliceBlue" Click="sideBarHelp_Click">
+						<TextBlock Text="Справка" TextWrapping="Wrap" TextAlignment="Center" FontSize="16" />
+						<Button.Resources>
+							<Style TargetType="Border">
+								<Setter Property="CornerRadius" Value="10"></Setter>
+							</Style>
+						</Button.Resources>
+					</Button>
+				</Grid>
+			</StackPanel>
+		</Grid>
+		<!-- Frame to display content -->
+		<Frame x:Name="contentFrame" NavigationUIVisibility="Hidden" Grid.Column="1" Margin="10" />
+	</Grid>
+</Window>
