+++ CourseLearning.Desktop/Models/Users.cs
@@ -1,0 +1,10 @@
+﻿namespace CourseLearning.Desktop.Models;
+
+public class User
+{
+    public int Id { get; set; }
+    public string Username { get; set; }
+    public string Password { get; set; }
+    public string FirstName { get; set; }
+    public string LastName { get; set; }
+}
