+++ CourseLearning.Desktop/Pages/HelpPage.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace CourseLearning.Desktop.Pages
+{
+    /// <summary>
+    /// Логика взаимодействия для HelpPage.xaml
+    /// </summary>
+    public partial class HelpPage : Page
+    {
+        public HelpPage()
+        {
+            InitializeComponent();
+        }
+    }
+}
