+++ CourseLearning.Desktop/Properties/Resources.resx
