+++ CourseLearning.Desktop/Properties/Settings.settings
