+++ CourseLearning.Desktop/Resources/Icons/CourseLearning.ico
