+++ CourseLearning.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CourseLearning.Desktop", "CourseLearning.Desktop\CourseLearning.Desktop.csproj", "{5766399A-C971-445B-A393-2DC18197D165}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5766399A-C971-445B-A393-2DC18197D165}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5766399A-C971-445B-A393-2DC18197D165}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5766399A-C971-445B-A393-2DC18197D165}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5766399A-C971-445B-A393-2DC18197D165}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
