--- CourseLearning_Lite.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34024.191
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CourseLearning_Lite", "CourseLearning_Lite.csproj", "{AD506518-C637-40E5-8C95-CF2AE2667359}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AD506518-C637-40E5-8C95-CF2AE2667359}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AD506518-C637-40E5-8C95-CF2AE2667359}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AD506518-C637-40E5-8C95-CF2AE2667359}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AD506518-C637-40E5-8C95-CF2AE2667359}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8DCE561A-E0C3-4543-8368-B443022EEDC8}
-	EndGlobalSection
-EndGlobal
