--- MainWindow.xaml
@@ -1,82 +1,0 @@
-﻿<Window x:Class="CourseLearning_Lite.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:CourseLearning_Lite"
-        mc:Ignorable="d"
-        MinHeight="600" MinWidth="1100" Height="600" Width="1100" 
-        WindowStyle="SingleBorderWindow" Background="AliceBlue" 
-        Title="CourseLearning_Lite" WindowStartupLocation="CenterScreen">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="Auto"/>
-            <ColumnDefinition />
-        </Grid.ColumnDefinitions>
-
-        <!-- Sidebar -->
-        <Grid Background="LightSkyBlue">
-            <StackPanel Grid.Column="0" Width="200" VerticalAlignment="Center">
-                <!-- <Button Content="Profile" Margin="10" Background="LightSkyBlue" /> -->
-
-
-
-            <Grid Margin="0 0 0 0">
-                    <Ellipse HorizontalAlignment="Left" Height="35" Margin="10,0,0,0" VerticalAlignment="Center" Width="35">
-                        <Ellipse.Fill>
-                            <ImageBrush ImageSource="/Images/icons8-writing-48.png" Stretch="Fill"/>
-                        </Ellipse.Fill>
-                    </Ellipse>
-                    <Button x:Name="sideBarCreatingCourses" Margin="50,10,0,10" HorizontalAlignment="Left" Width="140" 
-                    Background="AliceBlue" Click="sideBarCreatingCourses_Click">
-                        <TextBlock Text="Создание курсов" TextWrapping="Wrap" TextAlignment="Center" FontSize="16"/>
-                        <Button.Resources>
-                            <Style TargetType="Border">
-                                <Setter Property="CornerRadius" Value="10"></Setter>
-                            </Style>
-                        </Button.Resources>
-                    </Button>
-                </Grid>
-
-
-
-                <Grid Margin="0 10 0 10">
-                    <Ellipse HorizontalAlignment="Left" Height="35" Margin="10,0,0,0" VerticalAlignment="Center" Width="35">
-                        <Ellipse.Fill>
-                            <ImageBrush ImageSource="/Images/icons8-courses-48.png" Stretch="Fill"/>
-                        </Ellipse.Fill>
-                    </Ellipse>
-                    <Button x:Name="sideBarReadingCourses" Margin="50,10,0,10" HorizontalAlignment="Left" Width="140" 
-                    Background="AliceBlue" Click="sideBarReadingCourses_Click">
-                        <TextBlock Text="Прохождение курсов" TextWrapping="Wrap" TextAlignment="Center" FontSize="16"/>
-                        <Button.Resources>
-                            <Style TargetType="Border">
-                                <Setter Property="CornerRadius" Value="10"></Setter>
-                            </Style>
-                        </Button.Resources>
-                    </Button>
-                </Grid>
-
-                <Grid Margin="0 0 0 0">
-                    <Ellipse HorizontalAlignment="Left" Height="35" Margin="10,0,0,0" VerticalAlignment="Center" Width="35">
-                        <Ellipse.Fill>
-                            <ImageBrush ImageSource="/Images/icons8-help-48.png" Stretch="Fill"/>
-                        </Ellipse.Fill>
-                    </Ellipse>
-                    <Button x:Name="sideBarHelp" Margin="50,10,0,10" HorizontalAlignment="Left" Width="140" 
-                    Background="AliceBlue" Click="sideBarHelp_Click">
-                        <TextBlock Text="Справка" TextWrapping="Wrap" TextAlignment="Center" FontSize="16"/>
-                        <Button.Resources>
-                            <Style TargetType="Border">
-                                <Setter Property="CornerRadius" Value="10"></Setter>
-                            </Style>
-                        </Button.Resources>
-                    </Button>
-                </Grid>
-
-            </StackPanel>
-        </Grid>
-            <!-- Frame to display content -->
-        <Frame x:Name="contentFrame" NavigationUIVisibility="Hidden" Grid.Column="1" Margin="10" />
-    </Grid>
-</Window>
