--- Pages/HelpPage.xaml
@@ -1,30 +1,0 @@
-﻿<Page x:Class="CourseLearning_Lite.Pages.HelpPage"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:CourseLearning_Lite.Pages"
-      mc:Ignorable="d" 
-      d:DesignHeight="450" d:DesignWidth="800"
-      Title="HelpPage" Background="AliceBlue">
-
-    <Grid>
-        <ScrollViewer VerticalScrollBarVisibility="Auto">
-            <StackPanel>
-                <TextBlock TextWrapping="Wrap" Margin="10" FontSize="28">
-                Добро пожаловать в конструктор обучающих курсов! Для того, чтобы создать курс высокого качества, пожалуйста, следуйте этим рекомендациям:
-                </TextBlock>
-                <TextBlock TextWrapping="Wrap" Margin="10" FontSize="22">
-                • Структурируйте свой курс в соответствии с темой: определите тему курса и разбейте ее на логические блоки. Каждый блок должен соответствовать одной странице.<LineBreak/>
-                • Пишите ясные заголовки: каждая страница должна иметь понятный заголовок, который отражает ее содержание. Постарайтесь кратко расписывать каждый заголовок.<LineBreak/>
-                • Теоретический материал: напишите текст для каждой страницы, который четко и ясно объясняет тему, избегая лишних слов и необходимой информации. Разбивайте текст на абзацы. Не допускайте, чтобы количество символов превышало 6000. Для переноса на новую строку используйте клавишу Enter, а также обращайте внимание на ошибки и опечатки, на которые указывает поле ввода.<LineBreak/>
-                • Создайте тесты: Каждый курс должен содержать как минимум 1/4 долю тест по отношению к количеству страниц. Каждая такая страница должна содержать тест с четырьмя вариантами ответа. Вопросы должны быть формулированы ясно и быть связанны с темой курса. Правильным ответом должен быть только один из предложенных вариантов.<LineBreak/>
-                • Добавьте вопросы с письменным ответом: добавьте на каждой третьей странице вопрос с предполагаемым письменным ответом, который позволит студентам проявить свои знания и понимание темы. Вопрос должен ссылаться на материал страницы. Ответ на вопрос должен быть максимально простым для написания и быть не больше двух слов. Обращайте внимание на регистр и орфографию. Внимательный читатель должен без проблем дать правильный ответ на ваш вопрос.<LineBreak/>
-                </TextBlock>
-                <TextBlock TextWrapping="Wrap" Margin="10" FontSize="22">
-                Мы надеемся, что эти рекомендации помогут вам создать высококачественный обучающий курс.
-                </TextBlock>
-            </StackPanel>
-        </ScrollViewer>
-    </Grid>
-</Page>
