+++ README.md
@@ -1,11 +1,29 @@
 # CourseLearning_Lite
-Десктопное приложение «CourseLearning» предназначено для создания и прохождения обучающих курсов. Это вторая вариация, в которой вырезана авторизация, регистрация профилей через базу данных PostgreSQL.   
-Готовый курс сохраняется в JSON файле и может быть открыт другим пользователем в этом же приложении.
 
-| Страница создания курса             |
-| ------------------------------------------- |
-| ![скриншот страницы создания курса](Images/github/App_create.png) |
-| Страница прохождения курса          |
-| ![скриншот страницы прохождения курса](Images/github/App_read.png) |
-| Справка                                     |
-| ![скриншот справки](Images/github/App_help.png) |
+Десктопное приложение "CourseLearning" предназначено для создания и прохождения обучающих курсов. Это вторая вариация, в которой вырезана авторизация, регистрация профилей через базу данных PostgreSQL. Готовый курс сохраняется в `JSON` файле и может быть открыт другим пользователем в этом же приложении.
+
+## Снимки экрана
+
+### Страница создания курса
+
+![скриншот страницы создания курса](.github/App_create.png)
+
+### Страница прохождения курса
+
+![скриншот страницы прохождения курса](.github/App_read.png)
+
+### Справка
+
+![скриншот справки](.github/App_help.png)
+
+## Сборка
+
+Для сборки проекта достаточно запустить команду `make`:
+
+```sh
+make
+```
+
+Исполняемый файл можно будет найти по пути:
+
+`./CourseLearning.Desktop/bin/Release/7.0-windows/CourseLearning.Desktop.exe`
